attempting to use typescript
