+++ package.json
@@ -37,6 +37,7 @@
   },
   "devDependencies": {
     "@tsconfig/node20": "^20.1.4",
+    "@types/express": "^4.17.0",
     "eslint": "^9.13.0",
     "nodemon": "^3.0.1",
     "prettier": "3.0.3",
